requirements.txt
appdirs==1.4.4
asgiref==3.7.2
async-generator==1.10
attrs==22.1.0
autopep8==2.0.0
beautifulsoup4==4.11.1
blinker==1.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
cowsay==6.1
cryptography==42.0.0
defusedxml==0.7.1
django-allauth==0.60.1
django-appconf==1.0.6
django-cors-headers==4.3.1
django-imagekit==5.0.0
django-taggit==5.0.1
exceptiongroup==1.0.4
Flask==3.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.3
mypy==1.8.0
mypy-extensions==1.0.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.2
pilkit==3.0
Pillow==10.1.0
pluggy==1.3.0
psycopg2==2.9.9
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.3
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
selenium==4.6.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.25
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.9.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
vboxapi==1.0
webdrivermanager==0.10.0
Werkzeug==3.0.1
wsproto==1.2.0
